+++ src/node_info.rs
@@ -15,10 +15,10 @@ use crate::node_type::NodeType;
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct NodeInfo {
     //node_idx: NodeIdx,   // Could never be 0, as 0 is reserved for None
-    parent_idx: NodeIdx,   // Parent node index, 0 for root
-    prev_sibling: NodeIdx, // previous sibling, or last child of parent
-    next_sibling: NodeIdx, // Could be next_sibling or the node following the parent
-    first_child: NodeIdx,  // First child of this node
+    pub(crate) parent_idx: NodeIdx, // Parent node index, 0 for root
+    prev_sibling: NodeIdx,          // previous sibling, or last child of parent
+    next_sibling: NodeIdx,          // Could be next_sibling or the node following the parent
+    first_child: NodeIdx,           // First child of this node
     node_type: NodeType,
 }